MAINT-543 - CMS Clear cache now uses moved folders
*Note:* This task really doesn't need to be in a separate repo like
this any more _but_ we should leave it here until we know for sure we
won't need it from other non-CMS repos. There is configuration needed in
the CMS repo to make this work which will be in the CMS companion PR. It
could be made more self-contained here but it's likely not worth it
right now.

---------------

The old approach of just trying to `rm -rf` the cache folder contents
has proven problematic for a while now in that if the folder was being
written to, this would fail. It often took _many_ tries to get it to
work and now that we're doing more in Pipelines for deploys, we can't
let it continue.

New Strategy
============

The flow for clearing the cache is as follows:

1. Remove the previously created "old" cache folder if it exists
    - We don't remove this _after- clearing the cache in case it's being
      written to and we end up in the same situation as before with
      failures due to activity. Instead we just leave it behind for next
      time.
2. Create the _new_ cache folder with `-new` appended to the end
3. Invoke the Deployer task `deploy:writable` to ensure permissions are
   correct on the new folder **before** putting it in place
    - See [Github function](https://github.com/deployphp/deployer/blob/05bc19c4e8063d7426f2e50ae7ba7debc196827c/recipe/deploy/writable.php)
    - This minimizes the chance for any access issues after putting this
      folder in place if we were to adjust this _after_ moving
4. Move the **current** CMS cache folder to one appended with `-old`
5. Move the pre-created `-new` folder to `silverstripe-cache`

This should ensure that we don't get any locks when trying to manipulate
the folders and it should ensure proper permissions are in line with
what Deployer expects using their own task.